Update sample applications.
