+++ templates/chat-simplified/javascript/server/infrastructure/index.js
@@ -18,7 +18,11 @@ const getInfrastructure = async function () {
   }
 
   return {
-    ask: {},
+    ask: {
+      viewStore: {
+        messages
+      }
+    },
     tell: {
       viewStore: {
         messages
