--- templates/chat-simplified/javascript/server/views/messages/initializer.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const initializer = {
-  storeType: 'infrastructure/inmemory',
-
-  async initialize () {
-    // Intentionally left blank.
-  }
-};
-
-module.exports = { initializer };
