--- templates/chat-simplified/typescript/server/views/messages/MessagesItem.ts
@@ -1,6 +1,0 @@
-export interface MessagesItem {
-  id: string;
-  text: string;
-  likes: number;
-  timestamp: number;
-}
