--- templates/chat-simplified/typescript/server/views/messages/projections/liked.ts
@@ -1,14 +1,0 @@
-import { LikedData } from '../../../domain/communication/message';
-import { MessagesItem } from '../MessagesItem';
-import { ProjectionHandler } from 'wolkenkit';
-
-export const liked: ProjectionHandler<MessagesItem[], LikedData> = {
-  selector: 'communication.message.liked',
-
-  async handle (messageItems, domainEvent): Promise<void> {
-    const messageItem = messageItems.find(
-      messageItem => messageItem.id === domainEvent.aggregateIdentifier.id);
-
-    messageItem.likes = domainEvent.data.likes;
-  }
-};
