+++ templates/chat/javascript/server/views/messages/index.js
@@ -1,16 +1,8 @@
 'use strict';
 
 const { all } = require('./queries/all');
-const { initializer } = require('./initializer');
-const { liked } = require('./projections/liked');
-const { sent } = require('./projections/sent');
 
 const messages = {
-  initializer,
-  projectionHandlers: {
-    sent,
-    liked
-  },
   queryHandlers: {
     all
   }
