+++ templates/chat/typescript/server/infrastructure/index.ts
@@ -5,10 +5,14 @@ import { setupInfrastructure } from './setupInfrastructure';
 import { AskInfrastructure, TellInfrastructure } from 'wolkenkit';
 
 export interface Infrastructure extends AskInfrastructure, TellInfrastructure {
-  ask: {};
+  ask: {
+    viewStore: {
+      messages: Collection<Message> | Message[];
+    }
+  };
   tell: {
     viewStore: {
-      messages: Collection<Message> | Message[]
+      messages: Collection<Message> | Message[];
     };
   };
 }
